services/registration_service/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1
PyJWT==2.7.0
python-consul